# Conflicts:
#	src/main/java/com/verygana2/controllers/AdLikeController.java
#	src/main/java/com/verygana2/controllers/PetGameConfigController.java
#	src/main/java/com/verygana2/mappers/ConsumerDetailsMapper.java
#	src/main/java/com/verygana2/services/AdLikeServiceImpl.java
#	src/main/java/com/verygana2/services/UserServiceImpl.java
#	src/main/java/com/verygana2/services/finance/KeyWalletServiceImpl.java
#	src/main/java/com/verygana2/services/interfaces/AdLikeService.java
#	src/main/java/com/verygana2/services/interfaces/finance/KeyWalletService.java
#	src/main/java/com/verygana2/services/marketplace/CopaymentServiceImpl.java
#	src/main/java/com/verygana2/services/referrals/ReferralServiceImpl.java
#	src/main/java/com/verygana2/services/surveys/RewardService.java
#	src/main/resources/application-prod.yml

+++ src/main/java/com/verygana2/dtos/raffle/responses/RaffleSummaryResponseDTO.java
@@ -23,8 +23,8 @@ public class RaffleSummaryResponseDTO {
     private ZonedDateTime startDate;
     private ZonedDateTime endDate;
     private ZonedDateTime drawDate;
-    private Long totalTicketsIssued;
-    private Long totalParticipants;
+    private Integer totalTicketsIssued;
+    private Integer totalParticipants;
     private Long prizeCount;
     private boolean requiresPet;
 }
